HotFix Task No.6

+++ 0x00-hello_world/6-size.c
@@ -14,7 +14,7 @@ int main(void)
 	long long int n;
 	float f;
 
-	printf("Size of a char: %lu.\n", sizeof(c));
+	printf("Size of a char: %lu.\n", (unsigned long)sizeof(c));
 	printf("Size of an int: %lu.\n", sizeof(i));
 	printf("Size of a long int: %lu.\n", sizeof(l));
 	printf("Size of a long long int: %lu.\n", sizeof(n));
